+++ package.json
@@ -1,9 +1,12 @@
 {
   "name": "todo-app-react",
+  "name": "my-app",
+  "homepage": "http://brophy2321.github.io/todo-app-react",
   "version": "0.0.1",
   "private": true,
   "devDependencies": {
     "react-scripts": "0.2.1"
+
   },
   "dependencies": {
     "react": "^15.2.1",
