+++ src/ToDo.js
@@ -1,20 +1,66 @@
 // Developer TODO: add App component
 import React, { Component } from 'react';
-import Task from './Task';
+
 
 
 
 class ToDo extends Component {
 
   render(){
+
     return(
 
     <div>
     <h1> React To Do</h1>
-      <Task />
+      <Task currentTask={this.props.currentTask}  />
     </div>
 
   )}
 }
+
+
+
+
+class Task extends Component {
+    constructor(props) {
+      super(props)
+      this.state = {
+        currentTasks: props.currentTasks
+      }
+    }
+      handleTaskInput(e) {
+          this.setState({
+            query: e.target.value
+          })
+      }
+
+      submitQuery(e){
+        e.preventDefault();
+      {this.state.currentTasks}
+      }
+  render () {
+
+
+    return (
+      <div>
+        <h3>Task List</h3>
+        <form onSubmit={(e) => this.submitQuery(e)}>
+        <input
+          onChange={(e) => this.handleTaskInput(e)}
+          value={this.state.newItem}
+          type="text"
+          placeholder="To Do" />
+          <button type="submit">Add Task</button>
+          </form>
+          <div>
+          {this.props.currentTasks}
+      </div>
+      </div>
+
+
+    )
+  }
+}
+
 
 export default ToDo;
